+++ schema/test/sql/src/sql.rs
@@ -27,6 +27,10 @@ define_fixture_store!(
 ///
 
 #[entity(source_key = "schema/test/sql/src/sql.rs::entity::1",
+    audit_timestamps(
+        created_at(source_key = "created_at", ident = "created_at"),
+        updated_at(source_key = "updated_at", ident = "updated_at")
+    ),
     store = "SqlTestStore",
     version = 1,
     pk(fields = ["id"]),
@@ -51,6 +55,10 @@ pub struct SqlTestUser {}
 ///
 
 #[entity(source_key = "schema/test/sql/src/sql.rs::entity::2",
+    audit_timestamps(
+        created_at(source_key = "created_at", ident = "created_at"),
+        updated_at(source_key = "updated_at", ident = "updated_at")
+    ),
     store = "SqlTestStore",
     version = 1,
     pk(fields = ["id"]),