+++ testing/macro-tests/src/test/enum_payload.rs
@@ -18,6 +18,10 @@ pub mod test {
     };
 
     #[entity(source_key = "testing/macro-tests/src/test/enum_payload.rs::entity::nested::1",
+    audit_timestamps(
+        created_at(source_key = "created_at", ident = "created_at"),
+        updated_at(source_key = "updated_at", ident = "updated_at")
+    ),
         store = "TestStore",
         version = 1,
         pk(fields = ["id"]),
@@ -65,6 +69,10 @@ pub mod test {
     pub struct Stage {}
 
     #[entity(source_key = "testing/macro-tests/src/test/enum_payload.rs::entity::nested::2",
+    audit_timestamps(
+        created_at(source_key = "created_at", ident = "created_at"),
+        updated_at(source_key = "updated_at", ident = "updated_at")
+    ),
         store = "TestStore",
         version = 1,
         pk(fields = ["id"]),