+++ testing/macro-tests/src/test/field_projection_order.rs
@@ -16,6 +16,10 @@ mod tests {
     };
 
     #[entity(source_key = "testing/macro-tests/src/test/field_projection_order.rs::entity::nested::1",
+    audit_timestamps(
+        created_at(source_key = "created_at", ident = "created_at"),
+        updated_at(source_key = "updated_at", ident = "updated_at")
+    ),
         store = "TestStore",
         version = 1,
         pk(fields = ["id"]),